refactor: 변환 함수 유틸리티 파일로 분리
