+++ src/contexts/note-editor-context.tsx
@@ -1,129 +1,8 @@
 import { getNote, writeNote } from "@apis/note/note";
 import { Editor } from "@tiptap/react";
 import { NoteContent, WriteNoteRequest } from "@typedefs";
+import { transformContentForApi, transformContentFromApi } from "@utils/note-content-transformer";
 import React, { createContext, useContext, useState, ReactNode, useCallback, useEffect } from "react";
-
-// 노드에서 텍스트 추출 헬퍼 함수
-const extractTextFromContent = (content?: NoteContent[]): string => {
-  if (!content) return "";
-  return content
-    .map((node) => {
-      if (node.type === "text") return node.text || "";
-      if (node.content) return extractTextFromContent(node.content);
-      return "";
-    })
-    .join("");
-};
-
-// 백엔드 API 형식을 TipTap JSON으로 역변환
-const transformContentFromApi = (content: NoteContent): NoteContent => {
-  // vocacard 역변환
-  if (content.type === "vocacard") {
-    const questionText = content.attrs?.question_front || "질문을 입력하세요";
-    const answerText = content.attrs?.answer?.[0] || "답변을 입력하세요";
-
-    return {
-      type: "vocacard",
-      attrs: { reversed: content.attrs?.reversed ?? false },
-      content: [
-        {
-          type: "question",
-          content: [{ type: "text", text: questionText }],
-        },
-        {
-          type: "answer",
-          content: [{ type: "text", text: answerText }],
-        },
-      ],
-    };
-  }
-
-  // blankcard 역변환
-  if (content.type === "blankcard") {
-    const prefixText = content.attrs?.question_front || "앞 문장을 입력하세요";
-    const blankText = content.attrs?.answer?.[0] || "정답";
-    const suffixText = content.attrs?.question_back || "뒷 문장을 입력하세요";
-
-    return {
-      type: "blankcard",
-      content: [
-        {
-          type: "prefix",
-          content: [{ type: "text", text: prefixText }],
-        },
-        {
-          type: "blank",
-          content: [{ type: "text", text: blankText }],
-        },
-        {
-          type: "suffix",
-          content: [{ type: "text", text: suffixText }],
-        },
-      ],
-    };
-  }
-
-  // 다른 노드는 content 배열만 재귀적으로 변환
-  if (content.content) {
-    return {
-      ...content,
-      content: content.content.map(transformContentFromApi),
-    };
-  }
-
-  return content;
-};
-
-// TipTap JSON을 백엔드 API 형식으로 변환
-const transformContentForApi = (content: NoteContent): NoteContent => {
-  // vocacard 변환
-  if (content.type === "vocacard") {
-    const questionNode = content.content?.find((c) => c.type === "question");
-    const answerNode = content.content?.find((c) => c.type === "answer");
-
-    const questionText = extractTextFromContent(questionNode?.content);
-    const answerText = extractTextFromContent(answerNode?.content);
-
-    return {
-      type: "vocacard",
-      attrs: {
-        ...content.attrs,
-        question_front: questionText,
-        answer: [answerText],
-      },
-    };
-  }
-
-  // blankcard 변환
-  if (content.type === "blankcard") {
-    const prefixNode = content.content?.find((c) => c.type === "prefix");
-    const blankNode = content.content?.find((c) => c.type === "blank");
-    const suffixNode = content.content?.find((c) => c.type === "suffix");
-
-    const prefixText = extractTextFromContent(prefixNode?.content);
-    const blankText = extractTextFromContent(blankNode?.content);
-    const suffixText = extractTextFromContent(suffixNode?.content);
-
-    return {
-      type: "blankcard",
-      attrs: {
-        question_front: prefixText,
-        question_back: suffixText,
-        answer: [blankText],
-      },
-    };
-  }
-
-  // 다른 노드는 content 배열만 재귀적으로 변환
-  if (content.content) {
-    return {
-      ...content,
-      content: content.content.map(transformContentForApi),
-    };
-  }
-
-  return content;
-};
 
 interface NoteEditorContextType {
   noteId: number;
@@ -213,14 +92,6 @@ export const NoteEditorProvider: React.FC<NoteEditorProviderProps> = ({ children
         mode: "standard",
         contents: transformedContent,
       };
-
-      // 디버깅: 전송되는 데이터 확인
-      // console.log("=== 노트 저장 요청 데이터 ===");
-      // console.log("noteId:", noteId);
-      // console.log("name:", request.name);
-      // console.log("mode:", request.mode);
-      // console.log("contents:", JSON.stringify(editorContent, null, 2));
-      // console.log("==============================");
 
       // TODO: 이미지 파일 처리 로직 추가 필요
       const response = await writeNote(request);
